requirements.txt
# AI-Powered Communication Assistant Dependencies

# Backend Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1

# Frontend Framework  
streamlit==1.28.1
plotly==5.17.0

# Database
#sqlite3==2.6.0  # Built into Python, but listing for clarity

# Data Processing
pandas==2.1.3
numpy==1.25.2

# AI/ML
openai==1.3.7
tiktoken==0.5.2
httpx==0.25.2
sentence-transformers==2.7.0
scikit-learn==1.3.2
huggingface_hub==0.20.3
transformers==4.36.2
torch==2.1.2

# Email Processing
#imaplib3==0.9.5
#email-mime-base==1.1.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Utilities
requests==2.31.0
beautifulsoup4==4.12.2
html2text==2020.1.16
python-dateutil==2.8.2

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0